refactor(movex-core-utils): :recycle: more movings into movex-store for the unneeded core utils
