--- libs/movex-core-util/src/lib/index.ts
@@ -1,13 +1,0 @@
-export * from './misc';
-export * from './domain';
-export * from './Observable';
-export * from './EventEmitter';
-export * from './ScketIOEmitter';
-export * from './Logsy';
-export * from './ResultError';
-export * from './action';
-export * from './reducer';
-export * from './core-types';
-export * from './public-types';
-export * from './checkedState';
-export * from './io';
