+++ libs/movex/src/index.ts
@@ -1,16 +1,7 @@
 export * from './lib';
-// export * from '../../movex-master/src/lib/MovexMasterServer';
-// export * from '../../movex-store/src/lib';
-// export * from '../../movex-core-util/src/lib/public-types';
-
-// @deprecated
-export * as Client from './lib/client';
-// export * as Master from '../../movex-master/src/lib';
 
 // Aliased
 export * as MovexClient from './lib/client';
-// export * as MovexMaster from '../../movex-master/src/lib';
-// export * as MovexMasterLocal from './lib/master-local';
 
 export {
   type ResourceIdentifier,
@@ -22,4 +13,4 @@ export {
   isResourceIdentifier as isRid,
   isResourceIdentifierOfType as isRidOfType,
   isSameResourceIdentifier as isSameRid,
-} from  'movex-core-util';
+} from 'movex-core-util';
